+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/explosionoverhaul/ITRExplosion.java
@@ -205,10 +205,10 @@ public class ITRExplosion extends Explosion {
 						d11 *= 0.2d;
 					d11 *= getKnockbackMultiplier(this.source);
 					d11 = Math.min(d11, 10f);
-					entity.setDeltaMovement(entity.getDeltaMovement().add(xDistance * d11, yDistance * d11, zDistance * d11));
+					entity.setDeltaMovement(entity.getDeltaMovement().add(xDistance * d11, Math.max(yDistance * d11, 0.1f * this.radius), zDistance * d11));
 					if (entity instanceof Player player) {
 						if (!player.isSpectator() && (!player.isCreative() || !player.getAbilities().flying)) {
-							this.getHitPlayers().put(player, new Vec3(xDistance * d11, yDistance * d11, zDistance * d11));
+							this.getHitPlayers().put(player, new Vec3(xDistance * d11, Math.max(yDistance * d11, 0.1f * this.radius), zDistance * d11));
 						}
 					}
 				}